Fix method reference usages from constructors
This patch fixes that method references do not have
usages when they are constructor arguments.

The method reference symbols are resolved in
SymbolDataOfMethodReferenceBuilder which was only
used in the visit method of TypeVisitorAdapter
 for MethodCallExpr.

This patch adaptes SymbolDataOfMethodReferenceBuilder
to support constructors and the visit method of
ObjectCreationExpr is also updated to use
this component.